refactor: decompose large UI files into focused modules
